refactor: Util에 해당하는 모듈 이동
