+++ core/Profile.py
@@ -1,7 +1,8 @@
-import discord, random, asyncio, datetime
-import data.Logs as logs
+import discord, datetime
 from discord.ext import commands
 from PIL import Image,ImageDraw,ImageFont
+
+import utils.Logs as logs
 
 class Profile(commands.Cog):
